Add model that combines Prophet with the Box-Cox transform
The Box-Cox generates a parameter "lambda" during training that needs to be passed to the forecasting stage so that the correct inverse can be applied. I've done this as self.lam but there might be a cleaner interface for it

I'm really sorry, but I have hardly tested this code at all because I'm a bit short on time. I know you wanted something this week so I thought I'd better commit and pull request now rather than when I've got everything 100% working